Wip: Refactor HomeViewModel to load events from API or local storage
This commit refactors the HomeViewModel to use an EventsManager class for loading events. The loadAllEvents function now accepts a mockData parameter and calls either the getAllEvent method of the manager or loads data locally based on this parameter. The allEvents property is now published and can be observed by views that use this ViewModel.