Wip: Add HomeViewModel and CardStructView
This commit adds a new @StateObject called HomeViewModel to the HomeView struct. It also modifies the body of the HomeView by adding a NavigationView, ScrollView, and ForEach loop that displays events in cards using a new reusable CardStructView. The loadData() function is used to load mock data for now. Finally, this commit introduces a generic version of CardStructView that takes any SwiftUI view as its content and applies it to a background rectangle with rounded corners.